add multithreaded occupancy batch creation using reader & worker processes
Adds a separate function to create occupancy batches, using three separate
threads. A reader to only read raw data, a worker to extract the pixel data
and add it to the corresponding septemClasses and the main thread, which
calculates the occupancy from it.

Also refactors the read_zero_suppressed_data_file into two separate
functions.