add run time calculation based on shutter times and scaling of batches
Adds the calculation of the run times based on the actual shutter
opening times, rather than beginning and ending of a run. This is
important for shutter times shorter than 1 second, because dead
time starts to become significant.

Also adds automatic scaling, which is applied to the occupancy data,
which takes into account batches, which are shorter than one hour
and automatically scales the occupancy data accordingly.