add temperature readout as automatic logging to run function
Adds temperature readout to the run function based on the
MCP2210, which is read out via USB. It's connected to two
MAX31865.
One single function to test the readout is implemented
CommandTempLoopReadout()
as well as automatic logging in the run function.
For that a whole bunch of files are added, including
the HID, MCP2210 base communication functions and
abstract initialization and communication functions
built on top of this (explicitly not done in an object
oriented way, for the readout is very simple).
A separate thread is run in the waitconditions function,
which performs the measurements and logging.

Also moved the currentDateAndTime() function from the
hvFadcManager to a helper_functions module, since it's
also needed for the temperature logging.