move temperature thread from producer::run to pc to fix non-joining thread
Moves the thread, which performs the temperature logging from the
Producer::Run() function into the PC::DoRun() function, in which
it can be properly handled. Hence, all threads can exit properly
and we don't run into non-joining threads.
For that needed to define member class of type std::atomic_bool
for logging.