add occupancy data reading sorted by inode and foundation for temps in header
Changes the way the data is read in case of creation of occupancy dumps,
so that we don't read in the order of event numbers, but rather in
terms of the inode number. This speeds up reading of data by a
factor of 10.

Also lays foundation for displaying the temperatures for both sensors
on the header text, as well as a change to the way we read all files
in the folder. In the future: instead of reading the whole folder all
the time, we will read the contents once and then have a watcher thread,
which checks for changes in the folder and adds the created files
that way.