convert TOS numbering scheme for chips from 1,2,... to 0,1,...
This commit converts the numbering scheme used by TOS from 1 indexed
to 0 indexed, as one would expect from a C++ project.

Note: Currently the firmware expects a numbering starting from 1,
therefore we take this into account when sending the chip numbers to
the FPGA (Communication2 and CommunicationReadSend).

Note2: everything SEEMS to be working, but so far not even remotely
everything has been tested!