fix CTPR for TOCalib and zero supp. ro, + THLscan shutter ranges
Fixes the CTPR value of the TO calibration and zero suppressed
readout, which needs to be shifted similarly to the non-zero
suppressed readout.

Additionally, fixes a small bug in the THLscan of the previous commit,
which accidentally handed the shutter_time for the shutter_range.

Finally quite a few syntax changes.