correct create_tos_commands string handling
Fixes the creationg and handling of the C strings in the
create_tos_commands() function. Previously auto for loops
went over for (auto cmd : <set>), while on some compilers
this works fine, the assignment of
cmds[index] = cmd.c_str()
will result in non-valid data being stored in cmds[index],
once we leave the scope of the for loop. Thus, we go over
the references to the HFM and TOSCommandSets and leave the
references to these elements in the created const char**
array. Since they are defined as
static const std::string<std::set>, this works fine.