implement first version of temperature safety checks
This implements a first (compiling) version of a temperature safety
check for the detector. It checks whether the temperatures on the
intermediate board and the septemboard are within bounds, which are
defined in HFM_config.ini (currently lower 0 deg, upper 60 deg).

This is implemented by handing an AtomicTemps (typedefed in
mcp2210/temp_helpers.hpp) as a reference to the temp_auslese.cpp
init_and_log_temp() function, which is called in a separate thread in
the PC::runFadc() / PC::runOTPX() functions.
This AtomicTemps is also given (however as a pointer) to the
waitconditions Producer class (class constructor complains about a
reference. don't want to deal with compilation problems atm). Here,
the temperatures are only checked at the same time the HV is checked.
If an out of bounds is detected, the run is stopped.

A pair of atomics is used, because this way the temperature logging
thread can safely write to the variable, while the Producer can read
it at the same time. However, since no Floats can be used as
atomics (technically not easy to implement?), the temperatures are
rounded to integers. Enough for safety checks.

TODO: create separate thread which always (once TOS is started, until
quits) checks the temperature and, if exceeding bounds, shuts down HV.