add fns to set FADC trigger levels in mV and enable 14 bit readout
Adds functions to set, print, read the FADC trigger threshold not only
in the hardware 'ticks' of the FADC, but rather simply as a trigger
threshold in mV.
Introduces several functions to accomplish this, including helper
function to calculate on from the other.

Note: default values read from config file are still given in FADC
ticks.

Second major aspect: enables setting of Mode Register, which contains
bit 1: 14-bit mode. If this bit is set, the FADC (should at least) not
use only 12 bit for readout (splitting output in range of -1 to +1 V
over 12 bit => 4096 FADC ticks), but rather 14, resulting in a range
of 0 - 16384 ticks.

Functions above deal with both possibilities. Default in HFM_settings
is currently 14 bit mode.