add a temperature watcher thread started upon creation of HFM
Adds a separate thread, which is started upon creation of the
hvFadcManager, which only checks the temperature of the detector. If
the temperature leaves certain bounds defined in the HFM_settings.ini,
the HV is ramped down, if currently on. Will print Warnings either
way (all the time, until HFM is shut off!).
Read description of BackgroundTempLoop() in hvFadcManager.cpp for more
information.

Also changes some naked pointers to references (loop_continue
variables regarding temperatures). Use std::ref(), then it's fine hand
a reference to a separate thread!