add necessary parts to compile TOS under Cygwin on win (TOS.pro)
Adds necessary headers etc. to compile TOS under cygwin on a Windows
machine. Needed, because Cygwin on one hand acts as a unix system, but
is still dependent on some windows libraries.

Includes a configuration for Cygwin on Windows. In the TOS project
file one can select, whether to compile for Cygwin on Windows or
Unix.