implement CheckOffset using full matrix readout
Implements the CheckOffset function using full matrix readout, via a
5x5 chess matrix.

Calculation is done by creating a function of the Frame class, which
can calculate the difference between a given FrameArray and its
own. This value is returned and the preload with the smallest number
of errors is printed.

Also finally implements a set preload function.