remove calculations of preload in SaveData fns
Remove the references to the calculation of the preload being added to
the pixel values in the SaveData functions dealing with full matrix
readout.
The error, which caused random (periodic) artifacts on some pixels in
case one used full matrix readout, was found to be caused by the
preload. In case the preload is actively used, the bits by which the
preload shifts the reading of the data, caused the same number of bits
being lost at the end of each network package.

This is fixed by including the preload in the TOF, as it is done in
case of the zero suppressed readout.

The preload is still needed in TOS to determine, which value is
correct!

A SetPreload function was finally added to TOS as well.