change THLScan such that previous THL values are written
Changes the THLScan function such that after performing a THL scan for
a single chip, the THL value that was previously used on the chip is
written back. This is done to make sure not the last THL value, which
was written during the scan is active. In some cases (due to the daisy
chain of the chips), this could cause problems, if a chip was on a
very noisy THL.