change Run Producer such that it stops on Communication2() error
Changes the Run Producer, such that the run is stopped, if an error is
encountered in a Communication2() call, via DataChipFPGA() or
DataFPGAPC(). Previously, only on CountingTime() or CountingTrigger()
a check was made whether the return values were fine.
Done via a check in Producer while loop after these calls, which
immediately set PC's RunIsRunning() to false and break from loop.

This includes some clean up of error messages under DEBUG==1 in
fpga.cpp regarding timeouts of FPGA communication on select.

Also changes fpga.hpp DataFPGAPC() template, such that the case in
which an error of Communication2() is returned, the return value is
not (300 + err_code) anymore, but rather -(300 + err_code) to clearer
separate this case from the case in which we return the hits value!