implement access of temps of BackgroundTempLoop() into runFADC()
Implements the interconnection between the BackgroundTempLoop() and
the runFADC() function. Temperatures are checked the moment the HFM is
started. A reference of AtomicTemps is handed to the temperature
measuring thread. The value of the temperatures are handed to a Temps
'_currentTemps' object, owned by the HFM (note: note a AtomicTemps
object, since we cannot hand an AtomicTemps to a function, since
atomics are not copy contructable!).
From here, we can call the CheckIfTempsGood(), which simply reads the
_currentTemps variable and compares with safe temperatures.

NOTE: this commit does NOT include a temp_log.txt in the run folder!