change DataChipFPGA from template -> fn, better error handl. in run
Changes the template <typename Ausgabe> int DataChipFPGA(Ausgabe aus)
function to a normal function. The way it was implemented before was
nonsensical. A type Ausgabe was received as the function
argument. Then this argument was used to assign the Communication2()
output to it (note that Communication2() returns an int, meaning
Ausgabe independent of data type would store an int anyways!). Given
that Ausgabe could not be a pointer anyways, the value of 'aus' was
simply overriden by the assignment of the Communication2() result. At
the same time (note again: no pointer, means 'aus' was copied by value
upon function call!) the variable, which was handed to DataChipFPGA()
as 'aus' is still unchanged.

In many places a variable 'result' or similar was handed to
DataChipFPGA() in different places like
pc->fpga->DataChipFPGA(result);
which provides no benefit at all. A result variable has to be created,
while that result will never be available in the calling function! A
local int variable inside DataChipFPGA() provides exactly the same.
A
result = pc->fpga->DataChipFPGA(result);
would be necessary for that functionality, which is obviously (in the
given context at least!) complete uncalled-for.

Additionally, a template function of the specified type:
template <typename T> int Fn(T t);
which internally returns the value given by t (after assignment) as
the return value of the function explicitly demands that the type be
an int. It cannot be anything else regardless!

Finally, this commit also changes the error handling in
waitconditions.cpp, i.e. during a run. Each function call, which
performs an ethernet command to the firmware will be checked for,
stopping the run in case of any errors.