implement error handling if temp log file cannot be opened for write
Implements an error handling upon opening the temp log file to check,
whether writing to it is even possible. Previously it failed silently,
which meant no output was written in case e.g. the folder in which the
file was to be placed did not exist or access rights were wrong.