fix important bug in waitconditions.cpp about FADC, scinti vals
Fixes an important bug in waitconditions.cpp, which caused the FADC
and scintillator trigger values of the previous FADC trigger event to
be written to all subsequent events, until the next FADC trigger.

The problem was the if condition, in which the fadcParams map was
handed from the Producer to the Consumer. The case fadcReadout ==
false was not catched at all, which meant that the Consumer wrote the
old values (since its fadcParams remained unchanged) to the next files
over and over again.