add --full_matrix option to display full matrix r.o. frames
Adds an option to the event display to not only display zero
suppressed frames, but also full matrix readout.

Note, that in this case we demand that the --single_chip option is
also set, since we do not perform any parsing of header
information (none available in full matrix readout!).