fix proper HV ramping of Septemboard by ramping slowly (const ratio)
Implements a HV ramping scheme by which we ramp up the channels very
slowly, keeping the ratio of the voltages at the same level as the
target voltages.

Currently 150 steps are hardcoded, this is to be put into the ini file
as well. In addition the shutdown procedure has been taken out of TOS
for now. The ramp up will be adapted to a ramp down in the future.